refactor(workflows): dedupe start validation into one ctx-injected module (#510)
index.ts carried the workflow-start validator THREE ways: a module-scope
copy reading the `pluginCtx` global (for the REST route closures) and an
activate-scope copy closing over `ctx` (for hooks + exec-tools), each with
its own getRuntimeAgentNames / collectNestedWorkflowIds /
validateWorkflowForStart / createValidatedInstance. The two had quietly
diverged — only the route copy applied the trailing "assignee is not a
known runtime agent" superset check.

Collapse both into lib/start-validation.ts, a single recursive validator
with the runtime-agent dependency injected as a `PluginContext | null`
(null → empty agent set, matching the old module-scope guard). Routes pass
the `pluginCtx` global; hooks and exec tools pass `ctx`. The unified
validator keeps the superset assignee check, so REST keeps its behavior and
the hook/exec-tool paths gain the same check — the intended tightening to
one strong validator (the #510 goal: eliminate the three-way divergence).

index.ts drops to a single `createValidatedInstance` import (the other
three functions are now internal to start-validation) and sheds its
now-unused `createInstance` import.

Regression: a new gate-hooks test asserts the createInstance HOOK rejects a
nested-workflow cycle — parity with the existing REST cycle test — proving
all three start paths share the recursive validator.

Verified: typecheck + eslint + workflows suite (420/0, +1) + full suite
(5124/0) + binary build + isolated-home boot smoke (POST /instances/start
with a bogus workflowId returns 400 from createValidatedInstance, not 5xx).
